requirements.txt
python-telegram-bot==21.6
requests==2.32.3
APScheduler==3.10.4
